// .old_project /packages/cyoda-ui-lib/src/types/types.ts
export interface ReportingInfoRow {
  columnName: string;
  columnPath: string;
  type: string;
  elementInfo?: ReportingInfoRow;
  declaredClass?: {
    abstract: boolean,
    class: string,
  };
  elementType?: ReportingInfoRow;
  clazzType?: string;
  subClasses?: Array<{
    abstract: boolean,
    class: string,
  }>;
  keyInfo?: string;
}